+++ inc/settings/wps-admin-typography.php
@@ -8,18 +8,14 @@
 /* Add fonts */
 $myfont = WpsGenerateThemeFonts::get_instance();
 
+
+/* Sans-Sherif */
+
 $myfont->add_font('Open Sans',
 	'sans-serif',
 	'https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700,800',
 	'font-weight: 300',
 	'font-weight: 600'
-);
-
-$myfont->add_font('Merriweather',
-	'serif',
-	'https://fonts.googleapis.com/css?family=Merriweather:300,400,700,900',
-	'font-weight: 300',
-	'font-weight: 900'
 );
 
 $myfont->add_font('Raleway',
@@ -53,6 +49,135 @@ $myfont->add_font('Source Sans Pro',
 $myfont->add_font('PT Sans',
 	'sans-serif',
 	'https://fonts.googleapis.com/css?family=PT+Sans:400,700',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Dosis',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Dosis:200,300,400,500,600,700,800',
+	'font-weight: 300',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Nunito',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Nunito:200,300,400,600,700,800,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Muli',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Muli:200,300,400,600,700,800,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Rubik',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Rubik:300,400,500,700,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Catamaran',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Cabin',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Cabin:400,400i,500,600,700',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Josefin Sans',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Josefin+Sans:300,400,600,700',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Titillium Web',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Titillium+Web:200,300,400,600,700,900',
+	'font-weight: 300',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Exo 2',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Exo+2:200,300,400,500,600,700,800,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Roboto Mono',
+	'sans-serif',
+	'https://fonts.googleapis.com/css?family=Roboto+Mono:100,300,400,500,700',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+/* Cursive */
+
+$myfont->add_font('Expletus Sans',
+	'cursive',
+	'https://fonts.googleapis.com/css?family=Expletus+Sans:400,400i,500,600,700,700i',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+/* Sherif */
+$myfont->add_font('Merriweather',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Merriweather:300,400,700,900',
+	'font-weight: 300',
+	'font-weight: 900'
+);
+
+$myfont->add_font('Playfair Display',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Playfair+Display:400,400i,700,700i,900,900i',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Josefin Slab',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Josefin+Slab:300,400,400i,600,600i,700,700i',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Trirong',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Trirong:100,200,300,400,400i,500,600,600i,700,800,900',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Cormorant Garamond',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Cormorant+Garamond:300,300i,400,400i,700,700i',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Cormorant',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Cormorant:300,400,400i,700,700i',
+	'font-weight: 400',
+	'font-weight: 700'
+);
+
+$myfont->add_font('Arima Madurai',
+	'serif',
+	'https://fonts.googleapis.com/css?family=Arima+Madurai:100,200,300,400,500,700,800,900',
 	'font-weight: 400',
 	'font-weight: 700'
 );